Move tests to x-pack
