--- src/platform/test/functional/apps/dashboard/background_search/entrypoint.ts
@@ -1,64 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-import type { FtrProviderContext } from '../../../ftr_provider_context';
-
-export default function ({ getService, getPageObjects }: FtrProviderContext) {
-  const dashboardName = 'dashboard with filter';
-
-  const { dashboard } = getPageObjects(['dashboard', 'common', 'timePicker']);
-  const kibanaServer = getService('kibanaServer');
-  const security = getService('security');
-  const searchSessions = getService('searchSessions');
-
-  describe('dashboard background search', function () {
-    before(async () => {
-      await kibanaServer.savedObjects.cleanStandardList();
-      await kibanaServer.importExport.load(
-        'src/platform/test/functional/fixtures/kbn_archiver/dashboard/current/kibana'
-      );
-      await security.testUser.setRoles(['kibana_admin', 'test_logstash_reader', 'animals']);
-      await kibanaServer.uiSettings.replace({
-        defaultIndex: '0bf35f60-3dc9-11e8-8660-4d65aa086b3c',
-      });
-    });
-
-    after(async () => {
-      await security.testUser.restoreDefaults();
-      await kibanaServer.savedObjects.cleanStandardList();
-    });
-
-    beforeEach(async () => {
-      await dashboard.navigateToApp();
-      await dashboard.preserveCrossAppState();
-    });
-
-    describe('when in edit mode', () => {
-      describe('when clicking the open background search flyout button', () => {
-        it('opens the background search flyout', async () => {
-          await dashboard.gotoDashboardEditMode(dashboardName);
-
-          await searchSessions.openFlyout();
-          await searchSessions.expectManagementTable();
-        });
-      });
-    });
-
-    describe('when in view mode', () => {
-      describe('when clicking the open background search flyout button', () => {
-        it('opens the background search flyout', async () => {
-          await dashboard.loadSavedDashboard(dashboardName);
-
-          await searchSessions.openFlyout();
-          await searchSessions.expectManagementTable();
-        });
-      });
-    });
-  });
-}
