+++ src/platform/test/functional/services/index.ts
@@ -55,7 +55,6 @@ import { DashboardSettingsProvider } from './dashboard/dashboard_settings';
 import { ESQLService } from './esql';
 import { DataViewsService } from './data_views';
 import { SynthtraceClientProvider } from './synthtrace/sythtrace';
-import { SearchSessionsService } from './search_sessions';
 
 export const services = {
   ...commonFunctionalServices,
@@ -99,5 +98,4 @@ export const services = {
   savedObjectsFinder: SavedObjectsFinderService,
   esql: ESQLService,
   synthtrace: SynthtraceClientProvider,
-  searchSessions: SearchSessionsService,
 };
